Move NodeSourceAction class
